Refactor directory structure for user API and user tests
